Car-rental-Service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==1.0.4
python-dotenv==1.0.0
python-dateutil==2.8.2
PyJWT==2.10.1
cryptography==45.0.6
